+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.4 - 2019-03-19
+### Changed
+- Include jvm-metrics.yml.sample in package
+
+## 1.0.3 - 2019-03-18
+### Added
+- Added remote option to use remote URL connections. Format: service:jmx:remoting-jmx://host:port
+
+## 1.0.2 - 2019-02-13
+### Added
+- Added SSL option to Jmx.Open
+
 ## 1.0.1 - 2019-02-04
 ### Fixed
 - Updated protocol version